pyproject.toml
project = { dependencies = [
    "pyaudio>=0.2.14",
    "pip>=24.0",
] }
[tool.poetry]
name = "paper2audio"
version = "0.1.0"
description = ""
authors = ["Chenghao Mou <mouchenghao@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
transformers = "^4.38.2"
pymupdf = "^1.23.26"
numpy = "^1.26.4"
sentencepiece = "^0.2.0"
opencv-python = "^4.9.0.80"
regex = "^2023.12.25"
pyaudio = "^0.2.14"
tts = "^0.22.0"
torch = "^2.2.1"
torchvision = "^0.17.1"
google-cloud-texttospeech = "^2.16.3"
pydub = "^0.25.1"
typer = "^0.9.0"
rich = "^13.7.1"
alive-progress = "^3.1.5"


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
